+++ packages/server/lib/modes/init/scaffold/args.ts
